# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Health Record API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface PatientProfile {
  id: string;
  fullName: string;
  email: string;
  /** @nullable */
  dateOfBirth: string | null;
  /** @nullable */
  bloodType: string | null;
  initials: string;
}

export type DashboardSummaryCounts = {
  medications: number;
  allergies: number;
  conditions: number;
  labs: number;
  vitals: number;
};

export type HealthAlertSeverity = typeof HealthAlertSeverity[keyof typeof HealthAlertSeverity];


export const HealthAlertSeverity = {
  info: 'info',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface HealthAlert {
  id: string;
  severity: HealthAlertSeverity;
  title: string;
  description: string;
  actionLabel: string;
}

export interface DashboardSummary {
  patient: PatientProfile;
  counts: DashboardSummaryCounts;
  alerts: HealthAlert[];
  /** @nullable */
  nextReminder: string | null;
}

export type TimelineEntryDomain = typeof TimelineEntryDomain[keyof typeof TimelineEntryDomain];


export const TimelineEntryDomain = {
  conditions: 'conditions',
  medications: 'medications',
  allergies: 'allergies',
  labs: 'labs',
  vitals: 'vitals',
  immunizations: 'immunizations',
} as const;

export interface TimelineEntry {
  id: string;
  domain: TimelineEntryDomain;
  title: string;
  detail: string;
  date: string;
  source: string;
  /** @nullable */
  status: string | null;
}

export interface Medication {
  id: string;
  name: string;
  /** @nullable */
  dosage: string | null;
  /** @nullable */
  frequency: string | null;
  /** @nullable */
  route: string | null;
  /** @nullable */
  prescriber: string | null;
  /** @nullable */
  startDate: string | null;
  /** @nullable */
  endDate: string | null;
  isActive: boolean;
  /** @nullable */
  notes: string | null;
}

export interface MedicationInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  dosage?: string | null;
  /** @nullable */
  frequency?: string | null;
  /** @nullable */
  route?: string | null;
  /** @nullable */
  prescriber?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  isActive: boolean;
  /** @nullable */
  notes?: string | null;
}

export interface MedicationUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  dosage?: string | null;
  /** @nullable */
  frequency?: string | null;
  /** @nullable */
  route?: string | null;
  /** @nullable */
  prescriber?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  isActive?: boolean;
  /** @nullable */
  notes?: string | null;
}

export type AllergySeverity = typeof AllergySeverity[keyof typeof AllergySeverity];


export const AllergySeverity = {
  mild: 'mild',
  moderate: 'moderate',
  severe: 'severe',
  life_threatening: 'life_threatening',
} as const;

export interface Allergy {
  id: string;
  substanceName: string;
  /** @nullable */
  reaction: string | null;
  severity: AllergySeverity;
  isActive: boolean;
}

export type AllergyInputSeverity = typeof AllergyInputSeverity[keyof typeof AllergyInputSeverity];


export const AllergyInputSeverity = {
  mild: 'mild',
  moderate: 'moderate',
  severe: 'severe',
  life_threatening: 'life_threatening',
} as const;

export interface AllergyInput {
  /** @minLength 1 */
  substanceName: string;
  /** @nullable */
  reaction?: string | null;
  severity: AllergyInputSeverity;
  isActive: boolean;
}

export type ConditionStatus = typeof ConditionStatus[keyof typeof ConditionStatus];


export const ConditionStatus = {
  active: 'active',
  resolved: 'resolved',
  chronic: 'chronic',
} as const;

export interface Condition {
  id: string;
  displayName: string;
  status: ConditionStatus;
  /** @nullable */
  onsetDate: string | null;
  /** @nullable */
  notes: string | null;
}

/**
 * @nullable
 */
export type LabResultFlag = typeof LabResultFlag[keyof typeof LabResultFlag] | null;


export const LabResultFlag = {
  normal: 'normal',
  high: 'high',
  low: 'low',
  critical: 'critical',
} as const;

export interface LabResult {
  id: string;
  testName: string;
  value: string;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  referenceRange: string | null;
  /** @nullable */
  flag: LabResultFlag;
  collectedAt: string;
}

export type VitalVitalType = typeof VitalVitalType[keyof typeof VitalVitalType];


export const VitalVitalType = {
  blood_pressure: 'blood_pressure',
  heart_rate: 'heart_rate',
  glucose: 'glucose',
  weight: 'weight',
  spo2: 'spo2',
  temperature: 'temperature',
  respiratory_rate: 'respiratory_rate',
} as const;

export type VitalSource = typeof VitalSource[keyof typeof VitalSource];


export const VitalSource = {
  manual: 'manual',
  wearable: 'wearable',
  device: 'device',
} as const;

export interface Vital {
  id: string;
  vitalType: VitalVitalType;
  value: string;
  /** @nullable */
  unit: string | null;
  recordedAt: string;
  source: VitalSource;
}

/**
 * Invalid input
 */
export type BadRequestResponse = Error;

/**
 * Record not found
 */
export type NotFoundResponse = Error;

export type LimitParameter = number;

export type GetTimelineParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: LimitParameter;
domain?: GetTimelineDomain;
};

export type GetTimelineDomain = typeof GetTimelineDomain[keyof typeof GetTimelineDomain];


export const GetTimelineDomain = {
  all: 'all',
  conditions: 'conditions',
  medications: 'medications',
  allergies: 'allergies',
  labs: 'labs',
  vitals: 'vitals',
  immunizations: 'immunizations',
} as const;