# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Health Record API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryCounts } from './dashboardSummaryCounts';
import type { HealthAlert } from './healthAlert';
import type { PatientProfile } from './patientProfile';

export interface DashboardSummary {
  patient: PatientProfile;
  counts: DashboardSummaryCounts;
  alerts: HealthAlert[];
  /** @nullable */
  nextReminder: string | null;
}