# lib/api-zod/src/generated/types/healthAlert.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Health Record API
 * OpenAPI spec version: 0.1.0
 */
import type { HealthAlertSeverity } from './healthAlertSeverity';

export interface HealthAlert {
  id: string;
  severity: HealthAlertSeverity;
  title: string;
  description: string;
  actionLabel: string;
}