# lib/api-zod/src/generated/types/labResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Health Record API
 * OpenAPI spec version: 0.1.0
 */
import type { LabResultFlag } from './labResultFlag';

export interface LabResult {
  id: string;
  testName: string;
  value: string;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  referenceRange: string | null;
  /** @nullable */
  flag: LabResultFlag;
  collectedAt: Date;
}