# lib/api-zod/src/generated/types/medication.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Health Record API
 * OpenAPI spec version: 0.1.0
 */

export interface Medication {
  id: string;
  name: string;
  /** @nullable */
  dosage: string | null;
  /** @nullable */
  frequency: string | null;
  /** @nullable */
  route: string | null;
  /** @nullable */
  prescriber: string | null;
  /** @nullable */
  startDate: Date | null;
  /** @nullable */
  endDate: Date | null;
  isActive: boolean;
  /** @nullable */
  notes: string | null;
}