# lib/api-zod/src/generated/types/medicationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Health Record API
 * OpenAPI spec version: 0.1.0
 */

export interface MedicationInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  dosage?: string | null;
  /** @nullable */
  frequency?: string | null;
  /** @nullable */
  route?: string | null;
  /** @nullable */
  prescriber?: string | null;
  /** @nullable */
  startDate?: Date | null;
  /** @nullable */
  endDate?: Date | null;
  isActive: boolean;
  /** @nullable */
  notes?: string | null;
}